Hackwekend/GraphQL/Topic1/requirements.txt
flask
ariadne
uvicorn
python-dotenv